+++ pyworkflow/utils/__init__.py
@@ -33,3 +33,4 @@ from process import *
 from which import *
 from graph import *
 from reflection import *
+from log import *
